Commit: started on the player's house. It's about time this happened.

+++ Color.cs
@@ -112,7 +112,7 @@ namespace Noxico
 			return Color.FromArgb(r, g, b);
 		}
 
-		public static Color FromArgb(int argb)
+		public static Color FromArgb(long argb)
 		{
 			return new Color(argb, null);
 		}
